requirements.txt
streamlit==1.24.1
openai== 0.27.8
langchain==0.0.226
faiss-cpu==1.7.4
azure-storage-blob == 12.17.0
python-dotenv
psycopg2-binary==2.9.6
